work around bundler bug when constraining updates to a group
bundler has a bug when a gem is part of multiple groups

example Gemfile:
group :development, :test do
  gem ‘some_gem’
end
group :test do
  gem ‘another_gem’
end

when you run `bundle outdated -—group=test` it will only return info
with respect to `another_gem`, not `some_gem`

The bug is that the first group is registered as a group prefixed with
a space.  So to update all `test` group gems, you need to run the
outdated command twice, once without the space and a second time with
a space prefixed. Ugh!